python-services/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
open-rppg==0.1.4
opencv-python==4.10.0.84
numpy==1.26.4
